develop закончил написание тестов для функций в args.c. В test_args.md описал все тесты.

+++ tests/minunit.h
@@ -31,12 +31,12 @@ static const char *mu_current_test = "";
 
 #define mu_assert_inteq(expected, actual)                     \
     do {                                                      \
-        size_t e_ = (expected), a_ = (actual);                \
+        long e_ = (expected), a_ = (actual);                  \
         if (e_ != a_) {                                       \
             printf(                                           \
                 "FAIL %s (%s:%d): integers differ\n"          \
-                "    expected: %zu\n"                         \
-                "    actual:   %zu\n",                        \
+                "    expected: %ld\n"                         \
+                "    actual:   %ld\n",                        \
                 mu_current_test, __FILE__, __LINE__, e_, a_); \
             return 1;                                         \
         }                                                     \
